Allow to parse stored replication tasks using the "ref" key
Before change Ic0534d03dee6759ba4cb6b19ae87fd75269b437b a single
ref was stored per task using the "ref" key. The above change changed
it to the key name "refs", that had an array of refs as a value.
This led to issues, when tasks with the old format were still
present, when the change was applied.

Now, the old "ref" key is supported again and will be parsed correctly.

Bug: Issue 15922
Change-Id: I907109d0581ffa01781603741f4271471b19d71d