ReplicateRefUpdate: remove unnecessary factory method
The only difference in the signature of the two create methods was in
one parameter type: Set vs ImmutableSet. Since ImmutableSet is a subtype
of Set we can remove the method with the ImmutableSet in the parameter
type.

Note that ImmutableSet.copyOf(Collection c) is smart enough to not
create a copy of the passed collection when it is an ImmutableSet.

Change-Id: Ifc689027d17ff7ebdcab36b3f70a0adaf7942f09