Fix scheduling of tasks using the old storage format
The replication task storage uses the sha-sum of a string generated
from the objects properties as a filename. While scheduling the
task this shasum is computed again. Since the string which is the
basis for computing the shasum changed recently, this computation
results in a different result than the filename of the stored task.

Now the filename is used during scheduling and the shasum is not
recomputed. This allows to schedule tasks still using the old format.

Change-Id: I2dd0965ae0c7b1160f8c0774153f40edfae6c55a