Add allowManyToOneReplication option to remote configuration
Currently when there is the option of a replication configuration that
has multiple repositories, the URL is checked in order to verify that
the ${name} is present so that by mistake multiple repositories will not
override each other in target.

The protection is good for most of the cases.

Given an environment in which CI is triggered via a branch push event,
for example BitBucket, GitLab, GitHub. A simple integration with a CI
can be made by replicating the Gerrit changes into a branch at remote.
This will trigger CI per each Gerrit revision push.

In the current implementation, we can create 1:1 repositories Vat Gerrit
and remote for this purpose, with the following configuration:

    [remote]
        url = https://host.com/domain/${name}-ci.git
        push = +refs/changes/*:refs/heads/gerrit/changes/*
        replicationDelay = 2
        projects = domain/*

As changes are unique, there is no actual reason to open so many
repositories for this puprose... We can use a single repository for all,
with the following configuration:

    [remote]
        url = https://host.com/domain/common-ci.git
        push = +refs/changes/*:refs/heads/gerrit/changes/*
        allowManyToOneReplication = trie
        replicationDelay = 2
        projects = domain/*

We can even use the same repository with multiple Gerrit instances by
replacing the `gerrit` with environment identifier.

Of course, this is advanced feature and so noted in documentation,
however, it is very useful for integration with other environments.

Change-Id: I98565f6e683a5820233901ac4c8f4c972fd9e465
Signed-off-by: Alon Bar-Lev <alon.barlev@gmail.com>