Provide an option to skip replication of NoteDb meta refs
Replicating NoteDb meta refs is not needed when the remote does not run
a Gerrit instance.

Provide a config `remote.<NAME>.replicateNoteDbMetaRefs` which implies
whether to replicate the NoteDb meta refs (`refs/changes/*/meta`,
`refs/changes/*/robot-comments`, `refs/draft-comments/*`,
`refs/starred-changes/*`) or not. Time taken to replicate to a
non-Gerrit replica can be improved with this change.

Change-Id: I653f2e0a5b66dcded40c11ca23d2d22c63ccc396