Skip replicating NoteDb meta refs if remote does not run Gerrit instance
Replicating NoteDb meta refs is not needed when the remote does not run
a Gerrit instance.

Provide a config `remote.<NAME>.isGerritReplica` which implies whether
the remote replica runs a Gerrit instance or not. When set to false,
replication is skipped on NoteDb meta refs (`refs/changes/*/meta`,
`refs/changes/*/robot-comments`, `refs/draft-comments/*`,
`refs/starred-changes/*`). Time taken to replicate to a non-Gerrit
replica can be improved with this change.

Change-Id: I653f2e0a5b66dcded40c11ca23d2d22c63ccc396