PushOne: Do not assume URI is unique
Currently the tasks are grouped by URI of destination, this enforces the same
configuration to be used for all tasks.

The following configuration causes the replication to use the 1st fetched
configuration to replicate both remotes, resulting both project1 and project2
replicating into refs/heads/project1.

    [remote "project1"]
        url = https://host.com/domain/replica.git
        push = +refs/*:refs/heads/project1/*
        projects = project1
    [remote "project2"]
        url = https://host.com/domain/replica.git
        push = +refs/*:refs/heads/project2/*
        projects = project2

In order to to support a different RemoteConfig the task cache key should be
based on both the project (source) and the URI (destination).

Separate configuration per remote is also required for the upcoming
allowManyToOneReplication feature.

Change-Id: Ic0f905904af8800700d061063f85212dbeec58c2
Signed-off-by: Alon Bar-Lev <alon.barlev@gmail.com>