Add allowManyToOneReplication option to remote configuration
Currently when there is the option of a replication configuration that
has multiple repositories, the URL is checked in order to verify that
the ${name} is present so that by mistake multiple repositories will not
override each other in target.

The protection is good for most cases.

Given an environment in which CI is triggered via a branch push event,
for example BitBucket, GitLab, GitHub. A simple integration with a CI
can be made by replicating the Gerrit changes into a branch at remote.
This will trigger CI per each Gerrit revision push.

In the current implementation, we can create 1:1 repositories at Gerrit
and remote for this purpose, with the following configuration:

    [remote]
        url = https://host.com/domain/${name}-ci.git
        push = +refs/changes/*:refs/heads/gerrit/changes/*
        replicationDelay = 2
        projects = domain/*

There is no reason to enforce that so many repositories be used for this
purpose. We can use a single repository for all. As long as the ${name}
is included in the destination ref, so that any possibility of
conflicting refs can be avoided

For example the following configuration will push all changes from all
repositories as branches in a single repository to be used as a trigger
to a CI that monitors branch creation or modification.

    [remote]
        url = https://host.com/domain/common-ci.git
        push = +refs/changes/*:refs/heads/gerrit/${name}/changes/*
        allowManyToOneReplication = true
        replicationDelay = 2
        projects = domain/*

We can even use the same repository with multiple Gerrit instances by
replacing the `gerrit` with environment identifier.

Of course, this is advanced feature and so noted in documentation,
however, it is very useful for integration with other environments.

Change-Id: Ia25efa64578e2d5500f3679d12cdf20838a2d665
Signed-off-by: Alon Bar-Lev <alon.barlev@gmail.com>