Stop leaking replication task files on disk
When running Gerrit Replication with limitations configured (no
replication of config repos/refs, user ref visibility restrictions, etc)
this would lead to leaked replication tasks on disk in the waiting/
tasks directory.

The reason for this is that we applied a less stringent set of rules to
filter out projects/refs when writing the tasks to disk. Then when we
schedule the tasks we do further filtering. When this further filtering
occurs we may stop the process entirely or proceed with new replication
tasks with different ref lists. In the first case nothing follows up to
clean up the waiting/ tasks on disk they are simply orphaned. In the
second case the followup would fail to find matching files to clean up
because the details were different.

We fix this by applying all filtration of projects and refs before we
write these tasks to disk. This ensures that we'll either not write to
disk in the first place or run tasks that match what needs to be cleaned
up later.

Bug: Issue 40015481
Bug: Issue 40015482
Release-Notes: Fix 40015481 and 40015482 to stop leaking replication task files on disk
Change-Id: Ib70cb1d9f870b4dfb5fcdb04e467925c49d098da