Remove events from beforeStartupEventsQueue after fire()
The plugin was removing the event before firing it
from the beforeStartupEventsQueue. This could have
caused a message loss in case the configuration was
changed before firing the event.

Now after the plugin is fully loaded each event is
removed from the the in-memory queue after being fired.

Follow up of change 393036 .
Plese see comment [1]

[1]: https://gerrit-review.googlesource.com/c/plugins/replication/+/393036/comment/f9aed6f7_5c06661a/

Change-Id: I83abb264be3deae2f5b2330c70727e8f7b6cd3d3