Add `ConfigResource` interface
This is a first step for providing replication configuration out of the
server file system. The `ConfigResource` bundles together JGit `Config`
instance with its version.

The future implementations can read the configuration from a Gerrit
project or external systems like ZooKeeper.

Additionally, the `ConfigResource` separates the "configuration source"
from "configuration schema". The "schema" is provided by the
`ReplicationFileBasedConfig` that is hardcoding section and property
names used to get the configuration values from JGit `Config` object.
Whereas the `ConfigResource` interface is responsible for getting
the whole configuration object from somewhere. This could be from the
file system, a remote system or multiple files.

Change-Id: I77d205dce223e508bf4a06c5ff9da91444a87032