Convert `FanoutReplicationConfig` to `FanoutConfigResource`
With the introduction of `ConfigResource` we can simplify how the
"fanout" configuration works and reduce the code complexity.

The "fanout" replication configuration moves the `remotes` sections from
the `replication.config` file into the `replication/` directory with
one-remote-per-file approach.

What this effectively means, is that the JGit `Config` object provided
by with the "fanout" approach, is build not from a single file, but
potentially from multiple files.

This is one of the purposes of `ConfigReource`, to get the
configuration object regardless of how or where it's stored.

That's why, converting the `FanoutReplicationConfig` into
`FanoutConfigResource` makes sense. It also simplify the code, as
`@MainReplicationConfig` is no longer needed.

Change-Id: Ic6a5c5b8ab502a5e53d52a146d73b219563ee759