Add support for secure config information
Add support for the usage of the secure config plugin. The previous
implementation uses the secure.config with information in plain text,
which does not follow the expected pattern for this file. When using the
secure config plugin, the information now needs to be encripted. Notes are
added in the documentation regarding this change in behavior and how the
encripted information can be provided. In case the information in the
secure.config is not provided or is not encripted, logging information
is issued at the appropriate level.

Bug: Issue 320715545

Change-Id: Ibb29c6aa7cd82c95070286589d690bb0c6194f69