Use SecureStore instead of accessing secure.config
Implement SecureStore API in the access of secure information. The
previous implementation uses the secure.config directly, assuming that
relevant information is always in plain text, which does not always
happen for this file. When using the secure-config plugin,the
information needs to be encrypted.
Notes are added in the documentation regarding this change in behavior
and how the encrypted information can be provided. In case the
information provided by the SecureStore is not present or is not
encrypted, logging information is issued at the appropriate level.
Add replication.config username and password for situations where that
information must reside outside of the secure.config.

Bug: Issue 320715545
Change-Id: Ibb29c6aa7cd82c95070286589d690bb0c6194f69