Introduce `ReplicationConfigInvalidException` domain exception
When `ConfigParser` was introduced it relied on JGit's
`ConfigInvalidException` being thrown from its `parseRemotes` function
implementation.
The issue is that `ConfigInvalidException` (according to its description)
is meant to indicate that a given config is not a valid Git style
documentation. The `parseRemotes` receives `Config` as a parameter IOW
it is already ensured that it is valid from this perspective.

Introduce a dedicated `ReplicationConfigInvalidException` that should be
thrown when `Config` is not valid from the replication domain
perspective.

Change-Id: I46c0b3a4cf18f7e2311bcf4efc1c75b6227f505c