Update build to deploy Maven artefact
Since the addition of `ApiModule` to the plugin its API can be consumed
from other plugins. But to consume it the `replication` JAR file needs
to be added to the dependency list.

This works fine for Bazel-based in-tree plugins. For another type of
build systems we need to deploy the Maven artefact.

Recently the `gerrit_plugin` macro was updated to facilitate local
installation and deployment of Maven artefact. Now we, can add missing
parameter to deploy the `replication` artefact.

Change-Id: I304fcfd903754835c42771475c1a6afeda80712b