Remove DestinationConfigParser binding ReplicationConfigModule
The binding of the parsing logic of the replication config
is already in ReplicationModule L84 and should not be mixed
up with the implementation of how the config is stored.

Other replication plugins (e.g. pull-replication) use the
same replication config module for reading and storing the
replication.config but bind a different parsing logic
that would read the fetch sources and ignore the destination
targets.

This change is a pre-requisite for being able to fix the
Issue 330495476 that has caused the pull-replication to fail
on the master branch.

The origin of the breakage started with exposure of the
API to update the replication config introduced in I8003ec6
which exposed the internal implementation classes of the
replication plugin, instead of using the x-plugin communication
strategy that should only reply on interfaces exposed through
the ApiModule.

The design issue is easilyu resolved by exposing the updater
interface in the ApiModule and hiding the implementation from
the other plugins.

Also fix the code-style formatting of the imports for allowing
the successful validation of this change.

Bug: Issue 330495476
Change-Id: Ia5e6a5f3319687cbd4d3126f3b983d9266e4c065