Load replication configuration only once
Do not reload the configuration from disk continuously and reuse
the same instance across requests.

Depending on the configuration implementation (file-based, fanout),
the loading from storage and parsing in memory may require a
considerable amount of time, depending on the size of the config
and number of files of remotes endspoint.

Trigger the creation of a new Guice instance and the load from the
underlying storage may cause an unnecessary burden for the system.

The automatic reloading from disk is already implemented by a
specialised auto-reloader and regular injected instances are not
supposed to take changes form disk in real time.

Change-Id: I6f9c3a35d1ecbfa87f3daeeb464face0c6b6400e