Load replication configuration only once
Do not reload the configuration from disk
continuously and reuse the same instance across
requests in ReplicationConfigImpl.

Previously, the MergedConfigResource was stored as instance
variable in the ReplicationConfigImpl and the config was
retrieved on-demand with the Provider<>.get() paradigm.

However, the Provider<>.get() would trigger the creation
of a new instance of the Config every time is invoked,
triggering the full scan and parsing of the repication configs.

On GerritHub.io, the processing of all replication configs
(12k files) would take around 5 minutes, causing a catastrophic
overload of the CPU and preventing the plugin that use
ReplicationConfigImpl (replication and pull-replication) to
even start..

Change-Id: I6f9c3a35d1ecbfa87f3daeeb464face0c6b6400e