Reapply "Expose API to update remotes' configurations"
This reverts commit aab69373db6152ea98e7627f816fa3d777d5fe46.

As mentioned in the revered commit message, the plan was to only revert
that change in order to make 3.10 release. Now, when the release is
done we can get `ReplicatonRemotesUpdater` back.

Importantly, this reapply doesn't effectively bring back the
`ReplicationRemotesUpdater` API back in, as the definition of
`ApiModule` is still moved to `replication-api` artefact.

Change-Id: Id03d8d50cc662fca4af4c9a1a86d7ddb9f6254c0