Temporarily bind ReplicationRemotesUpdater as DynamicItem
To unblock other work around the github and pull_replication plugins,
bind the `ReplicationRemotesUpdater` as a `DynamicItem`. This allows us
to keep the API definition in the ApiModule and hide all of the
implementation details in the main replication plugin.

The `DynamicItem` is not an ideal choice here as it can be bound by any
plugin which could lead to unpredicted behaviour.

Hopefully, in the future, we'll be able to move to _StaticItem_ (Issue
338786480). Or refactor the plugin in a way that configuration details
could be moved into the `.api` package and stay hidden there.

For now, the best choice is to go with `DynamicItem`.

Change-Id: I349461e29d8cbd7a6d5e10c1e07ae8b7451b9f4d