Delay the fetching of credentials until push time
The replication plugin can be configured to reload dynamically
the credentials when changed on the filesystem or in the SecureStore
backend.

The overall reloading mechanism worked as long as a new PushOne
object was created per replication task; however, when a push was
failing because of credentials issues, the PushOne object was
requeued without being recreated, causing a chicken & egg situation

Until the PushOne was retried, the old credentials were stored
in the instance and therefore any retry would have failed.
Even though the replication.config was configured for reloading
dynamically the credentials, any task stuck in retry would have
not been able to be made pass with the correct credentials.

Delay the creation of the credentials provider until push
time, so that even in case of retrying replication tasks
the credentials can be dynamically reloaded when changed
on the filesystem.

Change-Id: I2f069a8433651a0db2681b6675f392ff894f27fc