Fix PushOne toString() to show the right refs and hidden number
Include all refs without applying any limit when `maxRefsToShow`
is set to 0 and limit the displayed refs to the specified maximum
when `maxRefsToShow` is greater than 0. Previously, the method
always limited the number of refs to `maxRefsToShow`, and
`maxRefsToShow` was updated to the batch count when initially
set to 0.

Remove the calculation of hidden refs to avoid performance
issues with millions of refs.

Bug: Issue 375798125
Release-Notes: skip
Change-Id: I30aae0cebf4fe28309542808baeeb316e1e7fb52