Fix PushOne toString() to show the right refs and hidden number
This changes improves the logic for handling the `maxRefsToShow`
configuration. When `maxRefsToShow` is set to 0, the method now
includes all refs without applying any limit. If `maxRefsToShow` is
greater than 0, the method limits the displayed refs to the specified
maximum. Additionally, the calculation of hidden refs has been removed
to avoid performance issues in cases with millions of refs.

Bug: Issue 375798125
Release-Notes: skip
Change-Id: I30aae0cebf4fe28309542808baeeb316e1e7fb52