Use try-with-resource to close Repository in PushOne
PushOne used an instance variable to store and share
the Repository object across the intermediate invocations
of its internal push phases.

Using an instance variable for Repository is causing
flakiness as it may not be properly cleaned up as also
shown in the tests executions with reference counting
enforced checks introduced with I303706438.

Change-Id: Ibb2fe90ae714e0be3a4f272fa992a555e2bf145f