docs/config: document inverted matches cancel each others
Our config had two inverted matches to exclude projects:

  [remote "github"]
    projects = ^(?:(?!apps\\/android\\/).)*$
    projects = ^(?:(?!apps\\/ios\\/).)*$

I have been wondering why `apps/android/wikipedia` is excluded given it
matches the first regex. The reason is ReplicationFilter applies each
pattern and when one matches it accepts the project. In this case, the
project is matched by the second and is thus accepted.

Ref: https://phabricator.wikimedia.org/T392198
Change-Id: I2846a40d2aabe80c0db305dc4f64dd1aa885b07d