Add explicit project exclusion support
While it was previously possible to exclude projects with the 'projects'
keyword by creating a negative match regular expression, this required
all the excluded projects to be specified in a single ORed regex which
is very ugly and error prone. Instead, introduce an explicit keyword,
'excludeProjects', which may be specified more than once to handle
project exclusions in a straight forward manner.

Release-Notes: Add explicit project exclusion support
Change-Id: I4e50f923f734537bc32b6a350d3268dbc286d818