Introduce an educated guess comment on why reflog is disabled on replicas
When repositories are replicated, their reflog is turned off on the
replica side. Add a comment explaining why that chose was possibly
made over a decade ago.

The author of the change is Shawn who passed years ago. The change
did not go through reviews at that time so there is no track record
of the choice.

Change-Id: I04de480df0d6012e3a4539fda945d15fc3f4eaa3